# Conflicts:
#	js/tempRoutes/jianyufeng.js
#	views/my/my.html

+++ controller/cart/cartController.js
@@ -97,8 +97,6 @@ define(['app',"./Fun/cart_fun"],function(app,cart_fun){
 		//递增按钮
 		cart_fun.addCartGoodsBtn(POP,function(countMoney,nowNum,gid,cartId,goodsNumber,limitNumber){
 
-			console.log(goodsNumber);
-			console.log(limitNumber);
 
 			var info = User.getInfo();
 			var updateParams = {
@@ -122,11 +120,6 @@ define(['app',"./Fun/cart_fun"],function(app,cart_fun){
 
 		//递减按钮
 		cart_fun.reduceCartGoodsBtn(POP,function(countMoney,nowNum,gid,cartId,goodsNumber,limitNumber){
-
-			console.log(goodsNumber);
-			console.log(limitNumber);
-
-
 
 			var info = User.getInfo();
 			var updateParams = {
@@ -180,18 +173,20 @@ define(['app',"./Fun/cart_fun"],function(app,cart_fun){
 
 		//点击删除
 		cart_fun.deleteCartBtn(function(_idx,_id){
-			
-			var info = User.getInfo();
-			var deleteParams = {
-				user_id : info.user_id,
-				shopping_type : 1,
-				id : _id
-			}
 
+			POP.Confirm("您是否放弃购买当前产品?",function(){
 
-			//删除购物车
-			
-			cartService.deleteCartGood($scope,deleteParams,POP,_idx,$rootScope);
+				var info = User.getInfo();
+				var deleteParams = {
+					user_id : info.user_id,
+					shopping_type : 1,
+					id : _id
+				}
+
+				//删除购物车
+				cartService.deleteCartGood($scope,deleteParams,POP,_idx,$rootScope);
+
+			});
 
 
 		})
